Consolidated quicksort into one directory
