+++ code/divide_conquer/quick_sort
@@ -1,0 +1,1 @@
+../sorting/quick_sort
