--- code/divide_conquer/quick_sort/README.md
@@ -1,3 +1,0 @@
-# Cosmos
-
-Collaborative effort by [OpenGenus](https://github.com/OpenGenus/cosmos)
