--- code/divide_conquer/quick_sort/quick_sort.hs
@@ -1,5 +1,0 @@
-quicksort :: (Ord a) => [a] -> [a]
-quicksort []        = []
-quicksort (x:xs)    = smaller ++ [x] ++ larger
-    where   smaller = quicksort $ filter (<=x) xs
-            larger  = quicksort $ filter (>x) xs
