+++ code/sorting/quick_sort/quick_sort.c
@@ -1,67 +1,75 @@
-/*Part of Cosmos by OpenGenus Foundation*/
-#include <stdio.h>
+/* Part of Cosmos by OpenGenus Foundation */
 
-void swap(int *p, int *q)
+/* C implementation QuickSort */
+#include<stdio.h>
+ 
+// A utility function to swap two elements
+void swap(int* a, int* b)
 {
-	int temp = *p;
-	*p = *q;
-	*q = temp;
+    int t = *a;
+    *a = *b;
+    *b = t;
 }
-
-
-//Last element is used a spivot
-//Places elements smaller than pivot to its left side
-//Places elements larger than pivot to its right side
-int partition(int a[], int low, int high)
+ 
+/* This function takes last element as pivot, places
+   the pivot element at its correct position in sorted
+    array, and places all smaller (smaller than pivot)
+   to left of pivot and all greater elements to right
+   of pivot */
+int partition (int arr[], int low, int high)
 {
-	int pivot = a[high];
-	int i = (low - 1);  
-
-	for (int j = low; j <= high- 1; j++)
-	{
-		if (a[j] <= pivot)
-		{
-			i++;
-			swap(&a[i], &a[j]);
-		}
-	}
-	swap(&a[i + 1], &a[high]);
-	return (i + 1);
+    int pivot = arr[high];    // pivot
+    int i = (low - 1);  // Index of smaller element
+ 
+    for (int j = low; j <= high- 1; j++)
+    {
+        // If current element is smaller than or
+        // equal to pivot
+        if (arr[j] <= pivot)
+        {
+            i++;    // increment index of smaller element
+            swap(&arr[i], &arr[j]);
+        }
+    }
+    swap(&arr[i + 1], &arr[high]);
+    return (i + 1);
 }
-
-void quickSort(int a[], int low, int high)
+ 
+/* The main function that implements QuickSort
+ arr[] --> Array to be sorted,
+  low  --> Starting index,
+  high  --> Ending index */
+void quickSort(int arr[], int low, int high)
 {
-	if (low < high)
-	{
-		int k = partition(a, low, high);
-		//k is pivot index
-
-		quickSort(a, low, k - 1);
-		quickSort(a, k + 1, high);
-	}
+    if (low < high)
+    {
+        /* pi is partitioning index, arr[p] is now
+           at right place */
+        int pi = partition(arr, low, high);
+ 
+        // Separately sort elements before
+        // partition and after partition
+        quickSort(arr, low, pi - 1);
+        quickSort(arr, pi + 1, high);
+    }
 }
-void print(int a[], int size)
+ 
+/* Function to print an array */
+void printArray(int arr[], int size)
 {
-	int i;
-	for (i = 0; i < size; i++)
-		printf("%d ", a[i]);
-	printf("\n");
+    int i;
+    for (i=0; i < size; i++)
+        printf("%d ", arr[i]);
+    printf("n");
 }
-
+ 
+// Driver program to test above functions
 int main()
-{   
-	int n, i;
-	printf("What is the size of the array?\n");
-	scanf("%d",&n);
-	int a[n];
-	printf("Enter elements of the array one by one\n");
-
-	for(i = 0; i < n; i++){
-		scanf("\n%d",&a[i]);
-	}    
-
-	quickSort(a, 0, n - 1);
-	printf("Sorted array: ");
-	print(a, n);
-	return 0;
+{
+    int arr[] = {10, 7, 8, 9, 1, 5};
+    int n = sizeof(arr)/sizeof(arr[0]);
+    quickSort(arr, 0, n-1);
+    printf("Sorted array: n");
+    printArray(arr, n);
+    return 0;
 }
